requirements.txt
Flask>=3.0
pymongo[srv]>=4.8
boto3>=1.34
python-dotenv>=1.0
Werkzeug>=3.0
gunicorn>=21.0
